added prizes to Ideas Elite page
We, Shrujan, Pranav, Yaswanthi and @Shantu0438 edited Ideas Elite webpage and added the prize details using col-md and col-lg. We used col-12 col-md-6 and col-lg-4 so that it looks good on phone as well. It was a wonderfully amazing experience. Thank You Pranav for teaching us what col-md and col-lg is. Belated Happy Teacher's Day Pranav !!! 🥳🥳🥳